chore: Add additional checks (#21)
* chore(misc): Update warp and warp-mp-ipfs

Refactor errors to add `Error::InvalidLength`, which would include context behind why the length is invalid

Added hard checks when updating the Identity. This may be customizable in the future

* chore(mp): Implemented a check before parsing stream

Note: This check is commented out for the time being.

* chore(mp): Use key directory when decrypting

* chore(rg): Increase internal time

* chore(rg): Added check for message length

Note: This will check all the lines for the length of the message, which should not exceed 4096 characters. This applies to new and edit messages.

* chore(misc): Added a check for raw data being transmitted

Currently, the limit in gossipsub for max transmitted data would be 256kb, and while its unlikely that any data sent would exceed this, we would want to check to be sure as the data is also encased in sata which would increase the size of the payload.

* chore(mp): Added an additional check against identity