feat(mp): Add `IdentityInformation` trait
This trait will contain 2 functions.

IdentityInformation::identity_status would get the status of an existing identity with the DID to determine of they are online or offline with `IdentityStatus`, which in the backend could be related to p2p connectivity, or some third party, database, etc. If a user is blocked, their status will always show offline.

IdentityInformation::identity_relationship would return `Relationship` that would contain flags of the relationship between the current identity and another existing identity. This would be related to friendship, if one received or sent a request to them, or has blocked said DID.