chore(mp): Switch to checking peer connection
Although this may change in the future or be a flag between different options, it may be better to check the connection of the peer rather than if they are available on a specific topic.

TODO: Add a toggle flag to allow using one or the other to determine online connectivity.