feat: Added additional functions to MultiPass and RayGun (#22)
* feat: Added additional trait functions in `MultiPass`

`Friends::received_friend_request_from` - This will allow one to determine if they received a request from an identity with the corresponding key

`Friends::sent_friend_request_to` - Will allow one to determine if a request was sent to an identity with a corresponding key

`Friends::is_blocked` - Will allow one to check if a specific key is blocked.

* feat(mp): Add `IdentityInformation` trait

This trait will contain 2 functions.

IdentityInformation::identity_status would get the status of an existing identity with the DID to determine of they are online or offline with `IdentityStatus`, which in the backend could be related to p2p connectivity, or some third party, database, etc. If a user is blocked, their status will always show offline.

IdentityInformation::identity_relationship would return `Relationship` that would contain flags of the relationship between the current identity and another existing identity. This would be related to friendship, if one received or sent a request to them, or has blocked said DID.

* feat: Added RayGun::get_message

This would allow one to get or obtain a single message from the conversation with the message id

* chore(mp): Implement trait functions

* add(rg): Implemented RayGun::get_message

* chore(mp): Switch to checking peer connection

Although this may change in the future or be a flag between different options, it may be better to check the connection of the peer rather than if they are available on a specific topic.

TODO: Add a toggle flag to allow using one or the other to determine online connectivity.

* chore(mp): Updated identity-interface example

* chore(mp): Implemented Display and FFIFree

* add(ffi): Implemented FFI functions

* chore(rg): Updated example

* chore(rg): Remove function from trait