refactor(config): Implemented `Discovery` into configuration
This would give options for different forms of discovery down the road.

Discovery::Provider would use the PUT_PROVIDER and GET_PROVIDER from libp2p to provide a context to DHT that would allow discovering of peers providing the same context.

Discovery::Direct will, in the future, would be used to query the DHT for the peer by using the peerid (which would derive from the public key, did, etc) to locate the peer that announces itself over DHT (this would require bootstrap flag to be true and using the same bootstrap nodes)

Discovery::None would disable discovery, which may be useful for testing over DHT, or using other forms of discovery in the future outside of DHT (eg blockchain, separate centralized server, etc).