add(misc): Implement traits against Arc<RwLock<Box<T>>>
With this implementation, we would be able to call the functions directly with it internally using the read or write lock. Downside is that the trait has to be imported for it to be recognized by the compiler.

Note: Constellation will not support this right now due to it containing references

Cherry picked from: ff7120413a759d431fe73d4f50a0163a5f9ea40f