feat/direct: Implement direct discovery
Discovery::Direct option will now allow ipfs search DHT for the peer by DID and connect to them upon discovering the information on it. Upon connecting, the identity information will be sent out. If we cannot find the peer the first around, we will schedule a background task to check for us.

Note: This may change to be its own future instead.