chore: Added a timeout and changes to function
We are doing this because the subscription registry in rust-ipfs causes some regression with specific task. In the case of a loop it gives kademlia enough time to query DHT then we timeout the function while the task completes in the background, although there is still a chance that we already completed it

Also added our own key to the discover_peer function so we can exclude it from the list if it shows up.