feat/direct: Implement direct discovery  (#36)
* feat/direct: Implement direct discovery

Discovery::Direct option will now allow ipfs search DHT for the peer by DID and connect to them upon discovering the information on it. Upon connecting, the identity information will be sent out. If we cannot find the peer the first around, we will schedule a background task to check for us.

Note: This may change to be its own future instead.

* chore/identity: use Ipfs::find_peer and add a type check

* add: Added function to check for connection

This function will check for a peer that is connected or subscribed to a specific topic

* chore: Use connected_to_peer to check if we are connected before query

* chore: Added a timeout and changes to function

We are doing this because the subscription registry in rust-ipfs causes some regression with specific task. In the case of a loop it gives kademlia enough time to query DHT then we timeout the function while the task completes in the background, although there is still a chance that we already completed it

Also added our own key to the discover_peer function so we can exclude it from the list if it shows up.

* add/queue: Added a placeholder for the queue

* chore/queue: Move logic over to the queue functionality

* chore/queue: Change logic to poll channel first and added `Queue::new`

* chore: Remove duration and added Queue::add_list and Queue::list

* chore: Implemented a small delay and check before proceeding with with the loop

* chore/request: Check connection directly in queue

Instead of checking if the peer we are looking for is subscribed to the topic, we should check for both if they are connected to us as well as subscribed to the topic.

* chore/queue: Added connection check to future

* refactor/connection: Change to use an enum to represent connection type

* chore: Clean up condition checks while polling

* chore: Move queue into its own module

* chore: Remove queue

This will be handled in a separate branch

* chore(mp): Use Ipfs::find_peer_info

* chore(mp): Added flag to identity-interface example

* chore: Use Ipfs::connected instead of Ipfs::peers

* chore: Added mdns flag to example

* chore: Update example comment

* chore/friends: Remove unnecessary functions

* chore(mp): Use default for development in configuration

* chore/tests: Added some basic test