feat: Implements Discovery::None
This option will disable discovery altogether and perform connections directly with each other. This would be establish through a common relay being used by peers (by default it will use the bootstrap as relays instead if they support the protocol). Bootstrap will also be disabled in the near future with this option.