add: Implemented PhoneBook
This will be used in conjunction with Discovery::Direct and Discovery::None to check to determine if the friend comes online or offline.

Note: Placeholder is in place to send an event out when a peer is detected to have gone offline or come online. This will likely be implemented in a later PR