feat: Implement Discovery::None (#40)
* feat: Implements Discovery::None

This option will disable discovery altogether and perform connections directly with each other. This would be establish through a common relay being used by peers (by default it will use the bootstrap as relays instead if they support the protocol). Bootstrap will also be disabled in the near future with this option.

* chore: Added flag check to persistent example

* add: Implemented PhoneBook

This will be used in conjunction with Discovery::Direct and Discovery::None to check to determine if the friend comes online or offline.

Note: Placeholder is in place to send an event out when a peer is detected to have gone offline or come online. This will likely be implemented in a later PR

* chore: Cleanup discovery_peer function

* chore: Updated identity-interface example

* chore: Dial and listen in on relays if they are in configuration

Although not used directly, but in the event that default bootstrap is not used where they are themselves a relay, using a custom relay list can be used.

* chore: Added additional profile to configuration

* chore: Use Into<PeerType> for connected_to_peer function