chore: Change lookup when checking for DID
This will check to see if the identity is cached and if not to check for a connection and pass discovery off if not connected.

Note: If Discovery::Provider is set, the discovery will return Ok because we are using DHT primarily. This will likely change to use DHT directly via find_peer