refactor: Move data into ipld (#64)
* refactor: Build ipld and send cid for identity instead of the identity object itself

* chore: Move logic from tokio::spawn into their own function

This would allow for better formatting and easier debugging

* chore: Remove pub keywords

* chore: Remove topic from connected_to_peer and change get_cid

* chore: Remove inline

* chore: Remove 2 fields from RootDocument

* fix: Continue iterating through the list if peer been seen

* chore: Remove duplicated event

* chore: Minor cleanup of code

* chore: Implement Hash for CacheDocument

* refactor: Store cache into ipfs

Note: This is a bit of a tricky part to do as we are hoping that we will be able to find the cid through bitswap. Since bitswap is 1.1 currently, this might be playing a part in some of the delays when using DHT.

* chore: Use path for log if available

* chore: Made changes to the caching

* chore: Use a reverted branch

Due to libp2p changes in 0.49, this may have had an impacted some code for bitswap.

* chore: Added conditional check when caching identity

* Revert "chore: Use a reverted branch"

This reverts commit 89ac5312431cdfaa2b0cee8b5daee5f8933f8e4a.

* chore: Increase the buffer size to max for mplex

* chore: Use a reverted branch

This reverts commit d542b096c6fce2a4dbbfae390e3ab248b0edf4c0.

* chore: Cleanup process_message and identity lookup

Make calls to the cache multiple times rather than a single time. This would allow us to be sure the cache is up to date.

* chore/clippy: Allow clone on copy.

* chore: Add functions to get and set the root document

* refactor: Move request, friends and block into ipld (WIP)

Note: Because of ipld using i128 for its integer, this would likely require mapping the request to a proper ipld format then storing it into a list.

* refactor: Have signatures be string instead of bytes

* fix: unpin block if its pinned

* refactor: Switch to using IdentityType

This type will either return an `Identity` or a `Cid` linking to the identity, which should allow compatibility while investigating bitswap.

* chore: Ignore clippy warning

* chore: Use DocumentType in place of IdentityType

This would allow switching between cid or a specific type if its being stored as a specific type or a cid linked to that type.

* chore/: Use DocumentType for other fields in `RootDocument`

TODO: Store objects in ipfs regardless of the type used

* fix: Use direct path instead of directory

* add: Implemented DID::from_str

* chore: Added check and minor cleanup

* fix: Load friends into phonebook upon initialization

* refactor: Use HashSet for storing request

* chore: Create list from an iterator

* chore: Added comment

* add: Added DocumentType::Resolve

This would resolve the object to the type either by returning `DocumentType::Object` or attempting to fetch it by using the cid with get_dag.

* chore: Use usable branch

* chore/identity-interface: Disable cache for the time being

* chore: Resolve the payload item from payload and cache

* chore: Cleanup and enable keepalive

* chore: Minor cleanup

* chore: Added a override config option

This would allow sending or using either the cid or the object itself. Use for compatibility after investigating more into bitswap to determine optimization that may be needed.

* chore: Use FuturesUnordered.

* chore: Update identity-interface example

Passthrough Opt to accounts and added override and context flags to cli.

Context flag would allow using a different provider context apart from default for discovering peers outside of the same provider key set

* add/document: Added additional resolve function to RootDocument and DocumentType

* chore: Minor cleanup and notations

Override ipld/dag by default - this will change back to false in the future
Make the override flag optional in the example

* chore: Update accounts test

* chore: Update configuration

* refactor: Check against pubsub peer list before sending

This will prevent the need of sending the identity every so often but instead would only send if there is a new peer who subscribed unless the identity updates then it would send out to everybody.

* chore: Added bootstap option to example

* chore: Remove unused calls

* chore: Change lookup when checking for DID

This will check to see if the identity is cached and if not to check for a connection and pass discovery off if not connected.

Note: If Discovery::Provider is set, the discovery will return Ok because we are using DHT primarily. This will likely change to use DHT directly via find_peer

* fix: Prevent recurring calls tokio::spawn in lookup

This will prevent recurring tokio::spawn by having a array of DID which has been called for discovery and will remove it once they are discovered

* chore: Remove initial check on cache list.