feat: Added fs-ipfs extension (#66)
* chore/rename: Rename warp-fs-ipfs to warp-fs-ipfs-client

* add: Initial implementation of warp-fs-ipfs

This implementation will use rust implementation of ipfs but will be using unixfs directly.

Note: This implementation will likely change in the near future to use [future] corresponding functions directly than to deal with unixfs

* chore/fs: Added example for fs-ipfs

* add: Implemented get_buffer and put_buffer

* add: Implemented Constellation::remove

* refactor: Make account optional

Although we still require multipass to be passed to the function, internally its optional in prep of separate initialization

* refactor: Rename a couple of errors

* fix: Use correct error

* chore: Implement `Hash`

* chore: Rearrange errors and added 2 additional errors

* feat: Implemented rename, create_directory. Added FFI constructor function

* chore/ipfs-example: Update example to use buffer instead

* chore: Implement SingleHandle::handle

* chore: Update build.rs

* add/c: Added small C example

* chore: Updated Makefile to include shared library

* chore: Temporarily enter the runtime context

This is mostly for shared library and being able to enter the tokio context for IO, spawning, etc. This doesnt have an impact on statically linked application.

* add: Added config

* chore: Exclude ffi function from cbindgen

* feat: Implemented config and import/export functions

Implementation of import and export will allow persistence with the index. On any mutable function it will attempt to export the index while also attempting to auto export in the event of any manual changes to the index

* chore: Update example

* chore: Added an additional example

* chore/ghaction: Update gh script