fix: Obtain permit when attempting to get a message
Calling get_message would likely cause a race condition making it possible that a message would not be obtainable after the event was emitted. To be sure that the message can be obtained after the event is emitted we could either emit the event after setting the root or obtain a permit from semaphore for that conversation id. This would make get_message wait its turn before calling any async task. This might change to were we will attempt to do this without obtaining a permit.