fix: Prevent blocking in example
Although in reality one wouldnt be pinning or unpinning every message (in fact there probably would be some sort of limit to pinning in a short amount of time), in the example when pinning a large amount of messages, it would stall and eventually panic or error due to everything being blocked. This will process everything in a separate task preventing it from blocking the loop.