refactor: Store messages in the blockstore (#104)
* chore: Move logic into separate functions

Since the IDE may not always detect things within macro, moving stuff into its own function would give more room to debug for errors

* chore: Use RwLock from tokio

* chore: Implemented Display for ConversationType

* add: Added DocumentType and Conversation documents

* chore: Cleanup future order list

* chore: Additional cleanup

* refactor: Remove DocumentType from MessageDocument and change logic

* chore: Add additional logic

* add: Added ConversationDocument::get_message

* chore: Move ConversationDocument into DocumentType

* chore: Added root_cid into DirectConversation

This would begin the process of moving the messages directly into ipfs blockstore

* add: Added functions to list and get conversation from document

* chore: Implement hash for DocumentType

* chore: Switch to BTreeMap instead of Vec for MessageDocument

* chore: Convert &CoversationDocument to Conversation

* chore: Migrate functions to use ConversationDocument

* refactor: Move ConversationDocument into its own module

* add: Implemented functions to build ConversationDocument

* chore: Refactor exist to use ConversationDocument

* chore: Additional changes to use ConversationDocument

* chore: Minor cleanup

* chore: Minor cleanup

* add: Added trait for conversion

Since TryFrom doesnt support async (yet) this will be used to handle conversion of any serializable type to DocumentType

* add: Implement a function to update the message in place

Will be useful for editing, pinning or reacting to a message without needing to deconstruct it each time

Note: This might be placed behind a guard to mutate in place.

* refactor: (WIP): Begin utilizing ConversationRootDocument

Tie functions into ConversationRootDocument, ConversationDocument and MessageDocuments

Implement Hash into ConversationDocument

Note: This will compile but will not work yet

* fix: Add event stream to map.

* chore: Update root document

* fix: Replace items in treeset after altering items

* chore: Added a task to hold permits

* chore: Added a map to hold tokio task and minor cleanup

This will allow for cleanup later on

* chore: Minor cleanup and update example

* chore: Use async-broadcast

* fix: Subscribe to topic after loading conversation

* chore/example: Disable events temporarily

Sending the events causes deadlocks and should be investigated

* fix: Dont compare messages

Because we are checking if they are eq already it would failed in cases of eg pinning a message

* fix: Obtain permit when attempting to get a message

Calling get_message would likely cause a race condition making it possible that a message would not be obtainable after the event was emitted. To be sure that the message can be obtained after the event is emitted we could either emit the event after setting the root or obtain a permit from semaphore for that conversation id. This would make get_message wait its turn before calling any async task. This might change to were we will attempt to do this without obtaining a permit.

* chore: Reimplement DeleteConversation

* chore: Minor cleanup

* fix: Workaround for deadlock when sending events that dont update document

TODO: Check to see if document needs updating before attempting to do so.

* chore: Acquire permit when updating document

* refactor: Remove permits and set conversation after processing message

* add/example: Implement function to get the number of messages in a conversation

* chore: Remove semaphore map

* chore: Change path

* refactor: Rename DirectMessageStore to MessageStore

* chore: Implement cleanup functions

* chore: added remove messagefunction in example

* fix: update root document after changes to message

* chore: Clean up function and minor change to function name

* chore: Switch back to tokio broadcast

* chore: Minor cleanup

* refactor: Use a struct instead of a enum for DocumentType

* refactor: Use Cid directly instead of DocumentType in ConversationDocument and MessageDocument

Although using DocumentType is still a option and doesnt hurt to use, the idea is to be able to walk down the tree manually easily without confusing fields

* chore: Cleanup logic

* fix: Update Message after initial reacting

* chore: Cleanup of reactions

* chore: Minor cleanup

* fix: Prevent blocking in example

Although in reality one wouldnt be pinning or unpinning every message (in fact there probably would be some sort of limit to pinning in a short amount of time), in the example when pinning a large amount of messages, it would stall and eventually panic or error due to everything being blocked. This will process everything in a separate task preventing it from blocking the loop.

* chore: Remove unnecessary configuration field

allowing unsigned message was used as a means of compatibility in previous version of this crate. Since a large time has past this field can be removed since it is no longer in use.

* chore: Remove lifetime

* chore: Implement conversation with reference

* add: Placeholders for future development

* chore: Cleanup and add logs

* chore: Use FuturesUnordered

* chore: Added a lock

* chore: Lower timeout

* refactor: Save conversations cid into their own files

Due to a race condition found under testing, we will proceed to save the cid outside of the single document for the time being and eventually merge them into a single document was a solution is implemented.

* fix: Remove conversation once deleted

* chore: Cleanup

* chore: Use reference

Although cloning is cheap for Ipfs<_>, borrowing it will keep it cleaner and friendlier while typing

* chore: use ReadDirStream to read directory

* fix: Remove block when replacing cid for conversation

* chore: Cleanup and optimization

* chore: Set permit limit to a const

This will allow adjusting the limit

* chore: Add some logging