chore: Revert the use of HashSet
The order of the hashset changes when deserializing it from ipld causing an invalid signature. Would have to look into that more since the hash should remain the same, however, the other option would be to use BTreeSet or use Vec, and for the time being we will revert to Vec and the possibility of using BtreeSet internally.