fix: Do not unpin and remove block if the cid is the same as the old
There may be times where the cid may be the same due to the function processing the event twice which would cause it to remove the block from the blockstore with the same cid. As a result would cause dag_get to await for finding a block if its not local. This small change would make it so that the block is only unpinned and removed if its not the same as the cid that was produced.