feat: Added base of group chat functionality (#111)
* feat: Added base of group chat functionality

* chore: Add variant to create group and edit example

* fix: Remove the additional recipient

* chore: Add signature placeholder

* feat: Sign the conversation document when its a group type

* feat: Add placeholder for adding and removing recipients

* refactor: Change recipient list to using HashSet

* fix: Verify only when group type is used

* chore: Cleaning with signing and verifying

* chore: Revert the use of HashSet

The order of the hashset changes when deserializing it from ipld causing an invalid signature. Would have to look into that more since the hash should remain the same, however, the other option would be to use BTreeSet or use Vec, and for the time being we will revert to Vec and the possibility of using BtreeSet internally.

* add: Implementation to add recipient to group conversation

* fix: Save conversation cid upon creation

* fix: Check to see if recipient exist before adding the list and signature

* chore: Implement function to remove recipient from group conversation

* chore: Updated example

* chore: Add a check to when conversation is being removed

This will make sure that if the event is emitted to other recipients that only the creator of the group can delete it.

* chore: Remove events from example

Although it does show it works, we dont need to send an event while sending a message

* chore: emit events when adding or removing recipient

* chore: Change order of event

* fix: Do not unpin and remove block if the cid is the same as the old

There may be times where the cid may be the same due to the function processing the event twice which would cause it to remove the block from the blockstore with the same cid. As a result would cause dag_get to await for finding a block if its not local. This small change would make it so that the block is only unpinned and removed if its not the same as the cid that was produced.

* chore: Minor update to example

* chore: Perform a check before removing block

* chore: Ignore message if sender not apart of conversation recipients

* chore: Minor cleanuo

* chore: Delete conversation for all recipients

Note: Only the creator of a group conversation will be able to delete the group for all recipients, while when a recipient deletes the group, it will only delete it for themselves

Note: Because the recipient list can only be updated by the creator, this will not remove the recipient from the list at this time, but rather in the near future when the creator receives an event from a recipient that they remove themselves that it would automatically run the function to alert the other recipients of updated list.