fix: Hold reference when getting conversation
Previously, we were getting a race condition due to cloning after obtaining the reference of the cid stored in memory, which before the PR works fine since we wanted the lock to drop, however this was seen to be problematic here because we were sending events to the task to be processed causing it to race ahead in an attempt to obtain the conversation again but would be invalidated after read. Dropping the clone and dereferencing the cid would allow us to hold into the lock until it drops at the end of the scope of the function.