refactor: Process event in same task (#120)
* refactor: Send events to the same task thats processing events from stream via channels

Originally we would use a semaphore to hold a place for events thats being process to prevent a race condition when putting a dag into ipfs block store and updating the cid held. This change will use channels to send an event and use the event from a tokio select to begin processing and storing. This, However, would make it so we would not need to use a semaphore exclusively for every single event but instead only use them when for the time being for adding or removing recipients from a conversation.

Note: Semaphore is still used for group conversation and relating to adding and removing a recipient but changes may be made to remove the need for them

* chore: Change get_conversation_mut to return a internal future

* Revert "chore: Change get_conversation_mut to return a internal future"

This reverts commit 70cb1e79fe249be4ee0aeae3e598683d2b640a33.

* chore: Cleanup

* fix: Hold reference when getting conversation

Previously, we were getting a race condition due to cloning after obtaining the reference of the cid stored in memory, which before the PR works fine since we wanted the lock to drop, however this was seen to be problematic here because we were sending events to the task to be processed causing it to race ahead in an attempt to obtain the conversation again but would be invalidated after read. Dropping the clone and dereferencing the cid would allow us to hold into the lock until it drops at the end of the scope of the function.

* chore: Add a optional oneshot channel to the conversation sender

This will allow us to return a `Result` if a channel is supplied to return an successful processing of the event or an error.

* chore: Remove unnecessary import

* chore: Drop conversation in event task

This is done so we dont hold anything additional in memory since we only need to grab the event topic associated with the conversation

* chore: Remove select_next_some

If the stream is polled after being close prematurely, it would cause a panic.