refactor: Move queue request into its own module (#147)
* refactor: Move queue into its own struct and task

This change will allow each entry to the queue to be checked in its own task rather than iterating over a list at a set interval.

* fix: Correct deadlock when closing request

When removing an item from queue and saving, it caused a deadlock due to us holding onto the write lock while using the if let condition. Separating it and then checking the variable would prevent us from holding onto the lock.