add: Implement listen for a response within a specific time frame.
This will use a oneshot to wait for a response when sending a request to determine if the user is blocked. The oneshot will time out specific a specific timeframe (as the time of this commit, its 500ms, but may increase in the future). If it receiver times out, it will proceed as if it was successful but if it the channel receives a result, it will try the result before proceeding.