feat: Change blocking functionality  (#148)
* feat: Add MultiPassEventKind::Blocked and MultiPassEventKind::Unblocked

* fix: Prevent blocking and unblocking one own key

* chore: Update example

* feat: Respond to request if user is block

This will notify the requester that they were blocked from that user and remove any request. This will also remove any outgoing and incoming request on the blocker side.

* chore: Add comments

* add: Implement listen for a response within a specific time frame.

This will use a oneshot to wait for a response when sending a request to determine if the user is blocked. The oneshot will time out specific a specific timeframe (as the time of this commit, its 500ms, but may increase in the future). If it receiver times out, it will proceed as if it was successful but if it the channel receives a result, it will try the result before proceeding.

* fix: Increase duration to 500ms

* chore: Add config flag for waiting for response

This will enables waiting for a response for a specific duration in milliseconds

* chore: Remove friend when block event received

When receiving a block event, we should remove them from the friend list as well, if are friends.

* add: Added block_by list

This will be used as a primary list to check who the user been blocked by before attempting send a request.

* chore: Update and check block_by list

To make sure we dont send needless request, we can check the list to determine who have blocked the user and update it when receiving events.

* add: Add blocked_by field to Relationship

* chore: Update test

* chore: broadcast event for closing or rejecting request when blocked