refactor: Switch to enum and support path
Previously, `IdentityUpdate` was an enum but due to wasmbindgen not supporting this variant of of enum (but instead only supported C style enums), it made it hard for compatibility, which lead it to be made as a struct. However, for the time being, it may be best to roll back the exporting of the functions to javascript (wrapping around wasm-bindgen) and allow for full wasm-bindgen usage. This will reduce complexity while allowing greater flexibility as it allows one to begin to write their own wrappers and avoid messy implementations

Along with this change also introduce and implements paths for pictures and banners, where we will attempt to store the images directly rather than storing it in a buffer.