feat(groups): Switch from open/closed concept to group settings
After some discussion, it seems that we don't want to have the concept
of open and closed groups but rather finer-grained indepdendant
settings. One example of a future setting would be "allow members to
edit group name". Once these changes are in, addition of such settings
would be easier in the future.